pyproject.toml
[project]
name = "interview-test-cake"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==3.1.5",
    "paramiko==3.5.1",
    "pendulum==3.1.0",
    "ruff>=0.14.10",
]
